Интерфейс, валидация полей, обработка ошибок и логика фильтрации записей по полям author, title и size - refactoring

+++ src/main/java/org/example/web/controllers/BookShelfController.java
@@ -4,6 +4,7 @@ import org.apache.log4j.Logger;
 import org.example.app.exceptions.IdNotFoundException;
 import org.example.app.services.BookService;
 import org.example.web.dto.Book;
+import org.example.web.dto.BookFilter;
 import org.example.web.dto.BookIdToRemove;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
@@ -15,11 +16,31 @@ import org.springframework.web.multipart.MultipartFile;
 
 import javax.validation.Valid;
 import java.io.*;
+import java.util.List;
 
 @Controller
 @RequestMapping(value = "/books")
 @Scope("singleton")
 public class BookShelfController {
+    @ModelAttribute("book")
+    public Book getBook() {
+        return new Book();
+    }
+
+    @ModelAttribute("bookFilter")
+    public BookFilter getBookFilter() {
+        return new BookFilter();
+    }
+
+    @ModelAttribute("bookIdToRemove")
+    public BookIdToRemove getBookIdToRemove() {
+        return new BookIdToRemove();
+    }
+
+    @ModelAttribute("bookList")
+    public List<Book> getBookList() {
+        return bookService.getAllBooks();
+    }
 
     private final Logger logger = Logger.getLogger(BookShelfController.class);
     private final BookService bookService;
@@ -32,9 +53,6 @@ public class BookShelfController {
     @GetMapping("/shelf")
     public String books(Model model) {
         logger.info("got book shelf");
-        model.addAttribute("book", new Book());
-        model.addAttribute("bookIdToRemove", new BookIdToRemove());
-        model.addAttribute("bookList", bookService.getAllBooks());
         return "book_shelf";
     }
 
@@ -42,8 +60,6 @@ public class BookShelfController {
     public String saveBook(@Valid Book book, BindingResult bindingResult, Model model) {
         if (bindingResult.hasErrors()) {
             model.addAttribute("book", book);
-            model.addAttribute("bookIdToRemove", new BookIdToRemove());
-            model.addAttribute("bookList", bookService.getAllBooks());
             return "book_shelf";
         } else {
             bookService.saveBook(book);
@@ -58,8 +74,6 @@ public class BookShelfController {
         logger.info("bookIdToRemove received: " + bookIdToRemove.getId());
         logger.info("Errors???: " + bindingResult.getAllErrors());
         if (bindingResult.hasErrors()) {
-            model.addAttribute("book", new Book());
-            model.addAttribute("bookList", bookService.getAllBooks());
             return "book_shelf";
         } else {
             bookService.removeBookById(bookIdToRemove.getId());
@@ -69,14 +83,18 @@ public class BookShelfController {
 
     @PostMapping("/removeByFilter")
     public String removeBookByFilter(
+            @Valid BookFilter bookFilter,
+            BindingResult bindingResult,
             Model model,
             @RequestParam String regAuthorToRemove,
             @RequestParam String regTitleToRemove,
             @RequestParam String regSizeToRemove) {
-        bookService.removeBookByFilter(regAuthorToRemove, regTitleToRemove, regSizeToRemove);
-        model.addAttribute("book", new Book());
-        model.addAttribute("bookList", bookService.getAllBooks());
-        model.addAttribute("bookIdToRemove", new BookIdToRemove());
+        model.addAttribute("bookFilter", bookFilter);
+        if (!bindingResult.hasErrors()) {
+            model.addAttribute("bookFilter", bookFilter);
+            bookService.removeBookByFilter(regAuthorToRemove, regTitleToRemove, regSizeToRemove);
+            return "redirect:/books/shelf";
+        }
         return "book_shelf";
     }
 
@@ -86,7 +104,6 @@ public class BookShelfController {
             @RequestParam String regAuthorToFilter,
             @RequestParam String regTitleToFilter,
             @RequestParam String regSizeToFilter) {
-        model.addAttribute("book", new Book());
         model.addAttribute("bookIdToRemove", new BookIdToRemove());
         model.addAttribute("bookList", bookService.filter(regAuthorToFilter, regTitleToFilter, regSizeToFilter));
         return "book_shelf";
@@ -94,7 +111,6 @@ public class BookShelfController {
 
     @PostMapping("/uploadFile")
     public String uploadFile(@RequestParam("file") MultipartFile file) throws IOException {
-
         String name = file.getOriginalFilename();
         byte[] bytes = file.getBytes();
 
@@ -116,19 +132,13 @@ public class BookShelfController {
     @ExceptionHandler(FileNotFoundException.class)
     public String fileUpload(Model model, FileNotFoundException exception) {
         model.addAttribute("fileNotFoundMessage", exception.getMessage());
-        model.addAttribute("bookIdToRemove", new BookIdToRemove());
-        model.addAttribute("book", new Book());
-        model.addAttribute("bookList", bookService.getAllBooks());
         return "book_shelf";
     }
 
     @ExceptionHandler(IdNotFoundException.class)
     public String idNotFound(Model model, IdNotFoundException exception) {
         model.addAttribute("idNotFoundMessage", exception.getMessage());
-        model.addAttribute("bookIdToRemove", new BookIdToRemove());
         model.addAttribute("bookIdToRemoveData", exception.getIdToRemove());
-        model.addAttribute("book", new Book());
-        model.addAttribute("bookList", bookService.getAllBooks());
         return "book_shelf";
     }
 }
